5.4 and 5.5 fixed.
In 5.4, when offline the bubble messages where displayed only in one side.
To fix 5.4, was created two new methods, saveUser() and getUser() where the user is saved together with the messages in AsyncStorage.
Afterwards getUser() is activated when Offline.
onCollectionUpdate - the _id is being fetch in 'doc' instead of 'data' to get the correct and fix the error [GiftedChat: cant find id in messages].